move flow editor to separate location
