+++ src/barfi/st_flow/base.py
@@ -54,7 +54,7 @@ def st_flow(
     commands: List[str] = ["execute", "save"],
     key=None,
 ):
-    base_blocks_data, _ = prepare_blocks_export(base_blocks)    
+    base_blocks_data, _ = prepare_blocks_export(base_blocks)
 
     # TODO: Add custom commands, to make it customize to act upon a command bar of tools
     # and commmands and how to render them on the ui
@@ -63,7 +63,7 @@ def st_flow(
         editor_schema=editor_schema,
         commands=commands,
         key=key,
-        default={"command": "default", "editor_state": {}},
+        default={"command": "default", "editor_schema": editor_schema},
     )
 
     return build_streamlit_flow_response(_from_client)
